requirements/base.txt
dj-database-url==0.4.1
django-debug-toolbar==1.5.0
django-extensions==1.7.4
Django==1.10.0
psycopg2==2.6.2
django-userena==2.0.1
django-extended-choices==1.1.1
